+++ capi/wrapper-generator/src/lib.rs
@@ -4,7 +4,7 @@
  * Created Date: 10/11/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 22/06/2023
+ * Last Modified: 28/07/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -52,6 +52,18 @@ pub fn generate<P: AsRef<Path>>(crate_path: P) -> Result<()> {
         &crate_path,
         true,
     )?;
+    gen::<CSharpGenerator, _, _>(
+        "../../dotnet/unity/Assets/autd3-linux/Scripts/NativeMethods",
+        &crate_path,
+        true,
+    )?;
+    if crate_path.as_ref().file_name().unwrap().to_str().unwrap() != "autd3capi-backend-cuda" {
+        gen::<CSharpGenerator, _, _>(
+            "../../dotnet/unity-mac/Assets/autd3/Scripts/NativeMethods",
+            &crate_path,
+            true,
+        )?;
+    }
 
     let out_file = Path::new("../../cpp/include/autd3/internal/native_methods").join(format!(
         "{}.h",
@@ -94,7 +106,6 @@ pub fn generate<P: AsRef<Path>>(crate_path: P) -> Result<()> {
                 ]
                 .into_iter()
                 .collect(),
-                
                 ..Default::default()
             },
             function: cbindgen::FunctionConfig {