+++ src/autd3/src/modulation/mod.rs
@@ -4,7 +4,7 @@
  * Created Date: 31/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 10/07/2023
+ * Last Modified: 28/07/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022-2023 Shun Suzuki. All rights reserved.
@@ -13,6 +13,7 @@
 
 mod cache;
 pub mod fir;
+pub mod fourier;
 pub mod radiation_pressure;
 pub mod sine;
 pub mod sine_legacy;
@@ -22,6 +23,7 @@ pub mod transform;
 
 pub use cache::Cache;
 pub use fir::FIR;
+pub use fourier::Fourier;
 pub use r#static::Static;
 pub use radiation_pressure::RadiationPressure;
 pub use sine::Sine;