+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Fix [#110](https://github.com/shinolab/autd3/issues/110): Multi-frequency sine modulation? 
 - Fix [#115](https://github.com/shinolab/autd3/issues/115): `autd3-geometry-viewer` from git does not work
 - Fix [#116](https://github.com/shinolab/autd3/issues/116): [autd3-unity] cannot launch simulator
+- Fix [#117](https://github.com/shinolab/autd3/issues/117): [autd3-unity] There is no LICENSE.md but LICENSE.txt
 
 # 14.1.0