+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 14.2.0
+
+- Add `modulation::Fourier`
+  - Fix [#110](https://github.com/shinolab/autd3/issues/110): Multi-frequency sine modulation? 
+- Fix [#111](https://github.com/shinolab/autd3/issues/111): Add macOS and Linux support for autd3-unity
+- Fix [#115](https://github.com/shinolab/autd3/issues/115): `autd3-geometry-viewer` from git does not work
+- Fix [#116](https://github.com/shinolab/autd3/issues/116): [autd3-unity] cannot launch simulator
+- Fix [#117](https://github.com/shinolab/autd3/issues/117): [autd3-unity] There is no LICENSE.md but LICENSE.txt
+
 # 14.1.0
 
 - Fix [#93](https://github.com/shinolab/autd3/issues/93): pyautd3 package contains unnecessary dynamic libraries